requirements.txt
openai
langchain
python-dotenv
requests
streamlit
tiktoken
langchain_community
langchainhub
PyGithub
agno
proxycurl
firecrawl-py
langchain-community
langgraph
streamlit-authenticator
supabase
pymupdf
SpeechRecognition 
pyaudio
gTTS